content_recommender/requirements.txt
flask>=2.3.2
numpy>=1.23.5
pandas>=1.5.3
faiss-cpu>=1.7.4
sentence-transformers==2.2.2
rapidfuzz>=3.13.0